fix server coverage by coverage_hooks.py
